fix(armv8): use GIC SPI priority regs to check priority bits
FreeRTOS code uses the write any, read legal mechanism to check which
priority bits in the GIC priority registers. By default it uses the
priority register of the first interrupt (an SGI). However, for GICv3
priority registers for SGIs and PPIs are read-as-zero in the distributor
as they are set in the redistributor. This commit changes this to use
the priority register of the first SPI, solving the issue.